requirements.txt
torch
torchvision
numpy
pillow
kornia